Development seta api ontology

+++ seta-data/src/app.py
@@ -1,18 +1,14 @@
+import json
+import os
 import time
 import shutil
-
-from elasticsearch import Elasticsearch
-#from sentence_transformers import SentenceTransformer
-from gensim.models import KeyedVectors
-
-from elasticsearch.helpers import bulk
+import time
+import config
 import requests
-import os
-from tqdm import tqdm
-import json
+from elasticsearch import Elasticsearch
+from elasticsearch.helpers import bulk
 
 import hashlib as hash
-
 from config import Config
 
 def getsha256(filename):
@@ -33,7 +29,7 @@ def wait_for_es():
         es_session = requests.Session()
         es_session.trust_env = False
         res = es_session.get(esh)
-        print(res,res.ok, flush=True)
+        print(res, res.ok, flush=True)
         if res.ok:
             res = json.loads(res.content)
             print("ElasticSearch...", res['status'], flush=True)
@@ -46,25 +42,19 @@ def wait_for_es():
         wait_for_es()
 
 
-def suggestion_update_job():
+def suggestion_update_job(config):
     models_path = Config.MODELS_PATH
-    if os.path.exists(models_path + Config.MODELS_WORD2VEC_FILE_CRC):
-        crc = open(models_path + Config.MODELS_WORD2VEC_FILE_CRC, 'r').read()
+    if os.path.exists(models_path + Config.WORD2VEC_JSON_EXPORT_CRC):
+        crc = open(models_path + Config.WORD2VEC_JSON_EXPORT_CRC, 'r').read()
     else:
-        crc = getsha256(models_path + Config.MODELS_WORD2VEC_FILE)
-        f = open(models_path + Config.MODELS_WORD2VEC_FILE_CRC,mode='w')
+        crc = getsha256(models_path + Config.WORD2VEC_JSON_EXPORT)
+        f = open(models_path + Config.WORD2VEC_JSON_EXPORT_CRC, mode='w')
         f.write(crc)
         f.close()
 
-    es_session = requests.Session()
-#    es_session.trust_env = True
-    
     index_suggestion = Config.INDEX_SUGGESTION
-    es = Elasticsearch("http://" + Config.ES_HOST, verify_certs=False, request_timeout=30)
-    resp = es_session.put(f"http://{Config.ES_HOST}/{index_suggestion}?pretty")      
-      
-    print(resp)
-    
+    es = Elasticsearch("http://" + config.ES_HOST, verify_certs=False, request_timeout=30)
+
     current_w2v_crc, crc_id = get_crc_from_es(es, index_suggestion)
     if crc != current_w2v_crc:
         try:
@@ -78,68 +68,78 @@ def suggestion_update_job():
                 es.update(index=index_suggestion, id=crc_id, doc=crc_model)
             else:
                 es.index(index=index_suggestion, document=crc_model)
-        except Exception as e:            
+        except Exception as e:
             print("errors on suggestion update. New crc: ", crc)
             print(e)
-            
+
 def gen_data(crc):
     print("suggestion indexing started")
-    
-    terms_model = KeyedVectors.load(Config.MODELS_PATH + Config.MODELS_WORD2VEC_FILE, mmap="r")
-    for word in tqdm(list(terms_model.wv.vocab)):
-        phrase = word.replace('_', ' ').lower()
+
+    with open(Config.MODELS_PATH + Config.WORD2VEC_JSON_EXPORT) as json_file:
+        data = json.load(json_file)
+
+    for suggestion in data:
         yield {
-            "_index": Config.INDEX_SUGGESTION,
-            "phrase": phrase,
-            "crc": crc
-        } 
-        
-def delete_all_suggestion(crc):    
-    print("suggestion delete started")
-    
-    query = {"query": {"bool": {"must": [{"match": {"crc.keyword": crc}}]}}}
-    es = Elasticsearch("http://" + Config.ES_HOST, verify_certs=False, request_timeout=30)
-    res = es.delete_by_query(index=Config.INDEX_SUGGESTION, body=query, wait_for_completion=False)
-    
-    print(res)
-    print("suggestion delete finished")
+                "_index": Config.INDEX_SUGGESTION,
+                "phrase": suggestion["phrase"],
+                "most_similar": suggestion["most_similar"],
+                "size": suggestion["size"],
+                "crc": crc
+            }
+
+def delete_all_suggestion(crc):
+    if crc:
+        print("suggestion delete started")
+        query = {"query": {"bool": {"must": [{"match": {"crc.keyword": crc}}]}}}
+        es = Elasticsearch("http://" + Config.ES_HOST, verify_certs=False, request_timeout=30)
+        res = es.delete_by_query(index=Config.INDEX_SUGGESTION, body=query, wait_for_completion=False)
+        print(res)
+        print("suggestion delete finished")
 
 
-def seta_es_init_map():
+def seta_es_init_map(config):
     es_session = requests.Session()
-#        es_session.trust_env = True
     headers = {"Content-Type": "application/json"}
     fn = Config.MODELS_PATH + Config.ES_INIT_DATA_CONFIG_FILE
-    
-    f = open(fn,'r')
-    dataformat = f.read()    
-    f.close()  
-    
-    print(dataformat)
+    f = open(fn, 'r')
+    data_format = f.read()
+    f.close()
     for indx in Config.INDEX:
-        resp = es_session.get(f"http://{Config.ES_HOST}/{indx}?pretty")
-        if resp.ok:
-            print("ElasticSearch index mapping exists: ",indx)
-        else:
-            resp = es_session.put(f"http://{Config.ES_HOST}/{indx}?pretty", data=dataformat, headers=headers)        
-            print (resp.content)
+        check_index_exists_or_create_it(Config.ES_HOST, data_format, es_session, headers, indx)
+    #suggestion index
+    fn = Config.MODELS_PATH + Config.ES_SUGGESTION_INIT_DATA_CONFIG_FILE
+    f = open(fn, 'r')
+    data_format = f.read()
+    f.close()
+    check_index_exists_or_create_it(Config.ES_HOST, data_format, es_session, headers, Config.INDEX_SUGGESTION)
+
+
+def check_index_exists_or_create_it(host, dataformat, es_session, headers, indx):
+    resp = es_session.get("http://" + host + "/" + indx + "?pretty")
+    if resp.ok:
+        print("ElasticSearch index mapping exists: ", indx)
+    else:
+        resp = es_session.put("http://" + host + "/" + indx + "?pretty", data=dataformat, headers=headers)
+        print(resp.content)
+
 
 def copy_models_files():
     dst = Config.MODELS_PATH + Config.ES_INIT_DATA_CONFIG_FILE
     src = Config.MODELS_DOCKER_PATH + Config.ES_INIT_DATA_CONFIG_FILE
     shutil.copyfile(src, dst)
 
-    dst = Config.MODELS_PATH + Config.MODELS_WORD2VEC_FILE
-    src = Config.MODELS_DOCKER_PATH + Config.MODELS_WORD2VEC_FILE
+    dst = Config.MODELS_PATH + Config.WORD2VEC_JSON_EXPORT
+    src = Config.MODELS_DOCKER_PATH + Config.WORD2VEC_JSON_EXPORT
     shutil.copyfile(src, dst)
 
-    dst = Config.MODELS_PATH + Config.MODELS_WORD2VEC_FILE + ".vectors.npy"
-    src = Config.MODELS_DOCKER_PATH + Config.MODELS_WORD2VEC_FILE + ".vectors.npy"
+    dst = Config.MODELS_PATH + Config.WORD2VEC_JSON_EXPORT_CRC
+    src = Config.MODELS_DOCKER_PATH + Config.WORD2VEC_JSON_EXPORT_CRC
     shutil.copyfile(src, dst)
 
-    dst = Config.MODELS_PATH + Config.MODELS_WORD2VEC_FILE_CRC
-    src = Config.MODELS_DOCKER_PATH + Config.MODELS_WORD2VEC_FILE_CRC
+    dst = Config.MODELS_PATH + Config.ES_SUGGESTION_INIT_DATA_CONFIG_FILE
+    src = Config.MODELS_DOCKER_PATH + Config.ES_SUGGESTION_INIT_DATA_CONFIG_FILE
     shutil.copyfile(src, dst)
+
 
 def get_crc_from_es(es, index_suggestion):
     query = {"bool": {"must": [{"exists": {"field": "crc_model"}}]}}
@@ -153,12 +153,21 @@ def get_crc_from_es(es, index_suggestion):
     return crc_es, crc_id
 
 
-           
+def getsha256(filename):
+    sha = hash.sha256()
+    with open(filename, 'rb') as f:
+        file_buffer = f.read(BLOCKSIZE)
+        while len(file_buffer) > 0:
+            sha.update(file_buffer)
+            file_buffer = f.read(BLOCKSIZE)
+    return sha.hexdigest()
+
+
 def init():
     wait_for_es()
     copy_models_files()
+    seta_es_init_map()
     suggestion_update_job()
-    seta_es_init_map()    
     print("SeTA-ES is initialised.")
